feat(custom-list): Save content when creating a new custom list
Previously, when creating a new custom list from the project screen, the content of the list was not saved. This was because the `onSaveCustomList` function in `ProjectScreenViewModel` did not pass the content to the `createCustomList` function in the repository.

This commit fixes the bug by passing the content to the repository.