feat(database): Migrate from Room to SQLDelight
This commit migrates the database from Room to SQLDelight.
The main reason for this migration is to have a single source of truth for the database schema, which is now defined in the `.sq` files.
This also solves the `table projects already exists` error, which was caused by the `onCreate` method being called multiple times.

The migration was done incrementally.
First, the `projects` table was migrated, and then the rest of the tables.
The `DatabaseModule.kt` was also updated to use the SQLDelight driver instead of the Room database.